feat: implement system to set field to NULL if empty - go back to system to set field like "or None" but now with   model_validator and pydantic - organize files better
